requirements/base.txt
pydantic>=2.6,<3
typer>=0.12,<1
sqlalchemy>=2.0,<3
psycopg[binary]>=3.1,<4
requests>=2.31,<3
python-dotenv>=1.0,<2
tenacity>=8.2,<9
